+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0).
 ### Added
 
 - `initialize` to set up the data directory (including a `readme.md` and the `search_details.csv`)
+- `merge_duplicates` with automated (threshold-based) and semi-automated identification of duplicates
 
 ### Changed